SPARQL queries
Hi @jonathanvajda, I created a fork for a version of the readme of project 4 that highlights the queries I am unsure about. After the initial part where I write down the toy rdf schemas, you can find a copy of the relevant exercises. Thanks in advance for your help and time.